requirements.txt
absl-py==0.12.0
atari-py==0.2.9
bsuite==0.3.5
dm-control==0.0.364896371
dm-env==1.5
dm-haiku==0.0.4
dm-launchpad==0.3.2
dm-reverb==0.4.0
dm-sonnet==2.0.0
dm-tree==0.1.6
flax==0.3.5
jax==0.2.19
jaxlib==0.1.70
keras==2.6.0
optax==0.0.9
Pillow==8.3.2
pytype==2021.8.11
pytest-xdist==2.3.0
rlax==0.0.4
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow==2.6.0
tfp-nightly==0.14.0.dev20210818
trfl==1.2.0
xmanager
